LayoutTests/fast/forms/file/entries-api/image-transcode-drag-drop-expected.txt

Tests a HEIC image will be converted to a PNG image when it is dragged and dropped to a file input element which accepts PNG.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS fileList.length is 1
PASS file.type is "image/png"
PASS file.size is non-zero.
Image was loaded.
PASS img.width is 400
PASS img.height is 400
PASS successfullyParsed is true

TEST COMPLETE
